requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
phonenumbers==9.0.19
pillow==12.0.0
sqlparse==0.5.3
